Persistence optimization (#593)
* optimizing

* Reducing the number of usage of whole save

* changing collectLastest to collect

* Changing all positions to double

* removing print

* calculating next and previous positions

* Changing position handling in WriteopiaStateManager

* Using next position in line break

* ordering

* Libe nreak working

* Fixing previous focus

* Update StepsModifier.kt

* Fisxing drag a feedback (with offset)

* Fixing offset

* Fixing space highlight

* fixing move

* removing whole

* Update OnUpdateDocumentTracker.kt

* fixing delete selection

* removing position normalization

* fixing tests

* build fix

* Update DocumentToMarkdownTest.kt

* Fixing tests

* code clean

* code clean

* Update DocumentRepositoryTest.kt

* Update DocumentsIntegrationTest.kt

* Update DragRowTarget.js.kt

* Update DragRowTarget.native.kt

* code clean

* Update NoteEditorScreen.native.kt

* Update SqlDelightDocumentRepositoryTest.kt

* Update init.sql

* Update DocumentSqlBeDao.kt

* changing queries

* Changing .sq

* Update build-and-test.yml

* Update build-and-test.yml

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Searching for error

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Update DocumentSqlBeDao.kt

* Update DocumentsRepository.kt

+++ application/core/documents/src/androidMain/kotlin/io/writeopia/core/folders/di/InDocumentSearchInjection.android.kt
@@ -13,7 +13,7 @@ actual class InDocumentSearchInjection {
         override suspend fun insertForFts(
             storyStep: StoryStep,
             documentId: String,
-            position: Int
+            position: Double
         ) {}
     }
