+++ plugins/writeopia_persistence_room/src/commonMain/kotlin/io/writeopia/sdk/persistence/entity/story/StoryStepEntity.kt
@@ -16,7 +16,7 @@ data class StoryStepEntity(
     @ColumnInfo(name = "path") val path: String? = null,
     @ColumnInfo(name = "text") val text: String? = null,
     @ColumnInfo(name = "checked") val checked: Boolean? = false,
-    @ColumnInfo(name = "position") val position: Int,
+    @ColumnInfo(name = "position") val position: Double,
     @ColumnInfo(name = "document_id") val documentId: String,
     @ColumnInfo(name = "is_group") val isGroup: Boolean,
     @ColumnInfo(name = "has_inner_steps") val hasInnerSteps: Boolean,
