+++ writeopia/src/commonMain/kotlin/io/writeopia/sdk/model/draganddrop/DropInfo.kt
@@ -1,3 +1,3 @@
 package io.writeopia.sdk.model.draganddrop
 
-data class DropInfo(val info: Any, val positionFrom: Int)
+data class DropInfo(val info: Any, val positionFrom: Double)
