+++ writeopia/src/commonMain/kotlin/io/writeopia/sdk/utils/alias/TypeAlias.kt
@@ -2,8 +2,8 @@ package io.writeopia.sdk.utils.alias
 
 import io.writeopia.sdk.models.story.StoryStep
 
-typealias DocumentContent = Map<Int, StoryStep>
+typealias DocumentContent = Map<Double, StoryStep>
 
-typealias UnitsNormalizationMap = (Map<Int, List<StoryStep>>) -> Map<Int, StoryStep>
+typealias UnitsNormalizationMap = (Map<Double, List<StoryStep>>) -> Map<Double, StoryStep>
 
-internal typealias UnitsMapTransformation = (Map<Int, StoryStep>) -> Map<Int, StoryStep>
+internal typealias UnitsMapTransformation = (Map<Double, StoryStep>) -> Map<Double, StoryStep>
