+++ writeopia/src/commonTest/kotlin/io/writeopia/utils/MapStoryData.kt
@@ -8,8 +8,8 @@ import io.writeopia.sdk.models.story.TagInfo
 
 object MapStoryData {
 
-    fun singleCheckItem(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun singleCheckItem(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             id = GenerateId.generate(),
             localId = GenerateId.generate(),
             type = StoryTypes.CHECK_ITEM.type,
@@ -17,8 +17,8 @@ object MapStoryData {
         )
     )
 
-    fun imageGroup(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun imageGroup(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.GROUP_IMAGE.type,
             steps = listOf(
@@ -39,29 +39,29 @@ object MapStoryData {
                 )
             )
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "4",
             type = StoryTypes.IMAGE.type,
         )
     )
 
-    fun imageStepsList(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun imageStepsList(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "1",
             type = StoryTypes.IMAGE.type,
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "2",
             type = StoryTypes.IMAGE.type,
         ),
-        2 to StoryStep(
+        2.0 to StoryStep(
             localId = "3",
             type = StoryTypes.IMAGE.type,
         ),
     )
 
-    fun imageSimpleGroup(): Map<Int, List<StoryStep>> = mapOf(
-        0 to listOf(
+    fun imageSimpleGroup(): Map<Double, List<StoryStep>> = mapOf(
+        0.0 to listOf(
             StoryStep(
                 localId = "1",
                 type = StoryTypes.IMAGE.type,
@@ -77,12 +77,12 @@ object MapStoryData {
         )
     )
 
-    fun twoGroupsImageList(): Map<Int, List<StoryStep>> {
+    fun twoGroupsImageList(): Map<Double, List<StoryStep>> {
         val parent1Id = GenerateId.generate()
         val parent2Id = GenerateId.generate()
 
         return mapOf(
-            0 to listOf(
+            0.0 to listOf(
                 StoryStep(
                     localId = parent1Id,
                     type = StoryTypes.GROUP_IMAGE.type,
@@ -129,8 +129,8 @@ object MapStoryData {
         )
     }
 
-    fun stepsList(): Map<Int, List<StoryStep>> = mapOf(
-        0 to listOf(
+    fun stepsList(): Map<Double, List<StoryStep>> = mapOf(
+        0.0 to listOf(
             StoryStep(
                 localId = "1",
                 type = StoryTypes.IMAGE.type,
@@ -144,13 +144,13 @@ object MapStoryData {
                 type = StoryTypes.IMAGE.type,
             )
         ),
-        1 to listOf(
+        1.0 to listOf(
             StoryStep(
                 localId = "4",
                 type = StoryTypes.TEXT.type,
             )
         ),
-        2 to listOf(
+        2.0 to listOf(
             StoryStep(
                 localId = "5",
                 type = StoryTypes.TEXT.type,
@@ -158,26 +158,26 @@ object MapStoryData {
         )
     )
 
-    fun singleMessage(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun singleMessage(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.TEXT.type,
             text = "hi!",
         )
     )
 
-    fun messagesInLine(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun messagesInLine(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.TEXT.type,
             text = "hi!",
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "2",
             type = StoryTypes.TEXT.type,
             text = "Hey!",
         ),
-        2 to StoryStep(
+        2.0 to StoryStep(
             localId = "4",
             type = StoryTypes.TEXT.type,
             text = """
@@ -186,7 +186,7 @@ object MapStoryData {
                 nisi ut aliquid ex ea commodi consequatur?
                 """,
         ),
-        3 to StoryStep(
+        3.0 to StoryStep(
             localId = "6",
             type = StoryTypes.TEXT.type,
             text = """I couldn't believe how sunny it was. Santiago is a really beautiful city.
@@ -194,14 +194,14 @@ object MapStoryData {
                 We went to many cozy Cafes and we enjoyed the city by foot.
                 We had to buy some Syn Cards to be able to communicate in the new country. """,
         ),
-        4 to StoryStep(
+        4.0 to StoryStep(
             localId = "7",
             type = StoryTypes.TEXT.type,
             text = "I couldn't believe how sunny it was. Santiago is a really beautiful city. There's a lot to " +
                 "do and enjoyed the day. We went to many cozy Cafes and we enjoyed the city by foot. " +
                 "We had to buy some Syn Cards to be able to communicate in the new country. ",
         ),
-        5 to StoryStep(
+        5.0 to StoryStep(
             localId = "8",
             type = StoryTypes.TEXT.type,
             text = "I couldn't believe how sunny it was. Santiago is a really beautiful city. There's a lot to " +
@@ -210,112 +210,112 @@ object MapStoryData {
         )
     )
 
-    fun simpleMessages(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun simpleMessages(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.TEXT.type,
             text = "message1",
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "2",
             type = StoryTypes.TEXT.type,
             text = "message2",
         ),
-        2 to StoryStep(
+        2.0 to StoryStep(
             localId = "4",
             type = StoryTypes.TEXT.type,
             text = "message3",
         ),
-        3 to StoryStep(
+        3.0 to StoryStep(
             localId = "6",
             type = StoryTypes.TEXT.type,
             text = "message4",
         ),
-        4 to StoryStep(
+        4.0 to StoryStep(
             localId = "7",
             type = StoryTypes.TEXT.type,
             text = "message5",
         ),
-        5 to StoryStep(
+        5.0 to StoryStep(
             localId = "8",
             type = StoryTypes.TEXT.type,
             text = "message6",
         )
     )
 
-    fun messagesWithHeader(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun messagesWithHeader(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.TEXT.type,
             text = "message1",
             tags = setOf(TagInfo(Tag.H1))
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "2",
             type = StoryTypes.TEXT.type,
             text = "message2",
             tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK, 0))
         ),
-        2 to StoryStep(
+        2.0 to StoryStep(
             localId = "4",
             type = StoryTypes.TEXT.type,
             text = "message3",
         ),
-        3 to StoryStep(
+        3.0 to StoryStep(
             localId = "6",
             type = StoryTypes.TEXT.type,
             text = "message4",
         ),
-        4 to StoryStep(
+        4.0 to StoryStep(
             localId = "7",
             type = StoryTypes.TEXT.type,
             text = "message5",
         ),
-        5 to StoryStep(
+        5.0 to StoryStep(
             localId = "8",
             type = StoryTypes.TEXT.type,
             text = "message6",
         )
     )
 
-    fun messagesWithHighlight(): Map<Int, StoryStep> = mapOf(
-        0 to StoryStep(
+    fun messagesWithHighlight(): Map<Double, StoryStep> = mapOf(
+        0.0 to StoryStep(
             localId = "0",
             type = StoryTypes.TEXT.type,
             text = "message1",
             tags = setOf(TagInfo(Tag.H1), TagInfo(Tag.HIGH_LIGHT_BLOCK, -1))
         ),
-        1 to StoryStep(
+        1.0 to StoryStep(
             localId = "2",
             type = StoryTypes.TEXT.type,
             text = "message2",
             tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK, 0))
         ),
-        2 to StoryStep(
+        2.0 to StoryStep(
             localId = "4",
             type = StoryTypes.TEXT.type,
             text = "message3",
             tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK, 1))
         ),
-        3 to StoryStep(
+        3.0 to StoryStep(
             localId = "6",
             type = StoryTypes.TEXT.type,
             text = "message4",
         ),
-        4 to StoryStep(
+        4.0 to StoryStep(
             localId = "7",
             type = StoryTypes.TEXT.type,
             text = "message5",
         ),
-        5 to StoryStep(
+        5.0 to StoryStep(
             localId = "8",
             type = StoryTypes.TEXT.type,
             text = "message6",
         )
     )
 
-    fun complexList(): Map<Int, List<StoryStep>> = mapOf(
-        0 to listOf(
+    fun complexList(): Map<Double, List<StoryStep>> = mapOf(
+        0.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.IMAGE.type,
@@ -329,7 +329,7 @@ object MapStoryData {
                     "hmac=xv-6mggpYPLIQ9eNAHrl1qKPHjyUCYlBoNBvdsqF4cY",
             ),
         ),
-        1 to listOf(
+        1.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.IMAGE.type,
@@ -337,7 +337,7 @@ object MapStoryData {
                     "hmac=xv-6mggpYPLIQ9eNAHrl1qKPHjyUCYlBoNBvdsqF4cY",
             )
         ),
-        2 to listOf(
+        2.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.IMAGE.type,
@@ -357,7 +357,7 @@ object MapStoryData {
                     "hmac=CaqZ-rcUAbmidwURZcBynO7aIAC-FaktVN7X8lIvlmE",
             )
         ),
-        3 to listOf(
+        3.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.TEXT.type,
@@ -369,7 +369,7 @@ object MapStoryData {
                      """,
             )
         ),
-        4 to listOf(
+        4.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.IMAGE.type,
@@ -378,7 +378,7 @@ object MapStoryData {
                 text = "The hotel entrance"
             )
         ),
-        5 to listOf(
+        5.0 to listOf(
             StoryStep(
                 localId = GenerateId.generate(),
                 type = StoryTypes.IMAGE.type,
@@ -389,7 +389,7 @@ object MapStoryData {
                     """,
             )
         ),
-        6 to listOf(
+        6.0 to listOf(
             StoryStep(
                 localId = "6",
                 type = StoryTypes.TEXT.type,
@@ -401,54 +401,54 @@ object MapStoryData {
         )
     )
 
-    fun syncHistory(): Map<Int, StoryStep> =
+    fun syncHistory(): Map<Double, StoryStep> =
         mapOf(
-            0 to
+            0.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.IMAGE.type,
                     url = "https://fastly.picsum.photos/id/15/400/400.jpg?" +
                         "hmac=xv-6mggpYPLIQ9eNAHrl1qKPHjyUCYlBoNBvdsqF4cY",
                 ),
-            1 to
+            1.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.IMAGE.type,
                     url = "https://fastly.picsum.photos/id/1018/400/400.jpg?" +
                         "hmac=MwHJoMaVXsBbqg-LFoDVL6P8TCDkSEikExptCkkHESQ",
                 ),
-            2 to
+            2.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.IMAGE.type,
                     url = "https://fastly.picsum.photos/id/984/400/400.jpg?" +
                         "hmac=CaqZ-rcUAbmidwURZcBynO7aIAC-FaktVN7X8lIvlmE",
                 ),
-            3 to
+            3.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.TEXT.type,
                     text = "We arrived in Santiago!!",
                 ),
-            4 to
+            4.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.CHECK_ITEM.type,
                     text = "We need to go to the Cafe!",
                 ),
-            5 to
+            5.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.CHECK_ITEM.type,
                     text = "We need to have lots of fun!",
                 ),
-            6 to
+            6.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.CHECK_ITEM.type,
                     text = "We need to lear some Spanish!",
                 ),
-            7 to
+            7.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.IMAGE.type,
@@ -456,7 +456,7 @@ object MapStoryData {
                         "hmac=gh5_PZFkQI74GShPTCJ_XP_EgN-X1O0OUP8tDlT7WkY",
                     text = "The hotel entrance"
                 ),
-            8 to
+            8.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.TEXT_BOX.type,
@@ -465,13 +465,13 @@ object MapStoryData {
                          exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid
                          ex ea commodi consequatur?""",
                 ),
-            9 to
+            9.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.TEXT_BOX.type,
                     text = "I hope to to it again some day..."
                 ),
-            10 to
+            10.0 to
                 StoryStep(
                     localId = GenerateId.generate(),
                     type = StoryTypes.TEXT_BOX.type,
